Move the testing of extra log elements
Moved from logging to testing, as we actually test the log testing /
validation functionality
